Remove BasicAer shot limit (#7801)
* Fixing issue #7634

At this point, there is an arbitrary number for the maximum shots in the BasicAer simulator "65536". The solution proposed @kevinsung is to work with zero as the default value of max_shots. This solves the problem with the evaluation of max_shots in the assemble function "_parse_common_args" in qiskit/compiler/assembler.py.

* Update backend.py

* Update backend.py

* Update backend.py

* adding test

correcting some errors about where to add the test.

* Create remove-basicaer-shot-limit-b7cd4e6f6739885c.yaml

* Update releasenotes/notes/remove-basicaer-shot-limit-b7cd4e6f6739885c.yaml

Co-authored-by: Kevin J. Sung <kevinjefferysung@gmail.com>

* Remove test of shot limit

The test previously just ran `BasicAer` with two million shots on the
transpiled circuit to see if it encountered a limit.  This is simple
enough functionality (and unlikely enough to be broken accidentally)
that it can do without a test - the previous one took about 8 seconds,
which is a long time for testing a toggle.

Co-authored-by: Kevin J. Sung <kevinjefferysung@gmail.com>
Co-authored-by: Jake Lishman <jake.lishman@ibm.com>