Add function to easily generate preset pass managers (#8125)
* Add function to easily generate preset pass managers

This commit adds a new function, generate_preset_pass_managers(), which
is used to quickly generate the preset pass manager for manual
interaction. Prior to this new function there were lower level functions
which are used to generate the individual pass manager objects for each
optimization level but they are hidden behind an abstraction object,
PassManagerConfig which is the only input. This makes it harder to work
with because the user has to manually remember to generate the config
object prior to creating the pass manager. Similarly this new function
simplifies the creation of preset pass managers by having a single entry
point instead of 4 for each optimization level.

* Add target override support if manually specified

If a user manually specifies a Target object to use as the compilation
target we still need to generate the old data structures to trigger
some passes to run. We also want the properties of the target to
supersede those in a specified backend if it's manually specified. This
commit corrects the handling of this so the manually specified target is
the source of truth when specified.

* Deduplicate shared argument list

* Expand test coverage

* Fix cyclic import issues

* Apply suggestions from code review

Co-authored-by: Julien Gacon <gaconju@gmail.com>

Co-authored-by: Julien Gacon <gaconju@gmail.com>